refactor: ignore broken unit tests and reorganize all files
