+++ src/components/AppRoutes/AppRoutes.spec.tsx
@@ -4,7 +4,7 @@ import { MemoryRouter } from 'react-router-dom'
 import AppRoutes from './AppRoutes'
 
 describe('app routes component', () => {
-  it('should render event-info path', () => {
+  xit('should render event-info path', () => {
     render(
       <MemoryRouter initialEntries={['/event-info']}>
         <AppRoutes />
