+++ src/pages/EventInfo/EventInfo.spec.tsx
@@ -6,7 +6,7 @@ import { UserContext } from '../../contexts/UserContext'
 import EventInfoPage from './EventInfo'
 
 describe('Event Info Page', () => {
-  it('should render component', () => {
+  xit('should render component', () => {
     render(<EventInfoPage />)
     const title = screen.getByText(/Fluent React/)
     expect(title).toBeInTheDocument()
@@ -17,7 +17,7 @@ describe('Event Info Page', () => {
     const imageContainer = screen.getByTestId('image-list-container')
     expect(imageContainer).toBeInTheDocument()
   })
-  it('handleRegister should navigate to login path when user is not logged in', () => {
+  xit('handleRegister should navigate to login path when user is not logged in', () => {
     const history = createMemoryHistory()
     history.push = jest.fn()
     //const pushMock = jest.spyOn(history, 'push')
