+++ src/shared/hooks/useAuth.tsx
